requirements.txt
aider-chat 
pydantic>=2.0.0
dependency-injector>=4.41.0
structlog>=23.1.0
rich>=13.3.5
tenacity>=8.2.2